==> analysis/07_cytomapper_images.Rmd <==
---
title: "plot images"
author: "Daniel"
date: "18.05.2020"
output: html_document
---

This script is used to start the cytomapper shiny app and perform the cell type labeling

```{r setup, include=FALSE}
knitr::opts_chunk$set(echo = TRUE)
knitr::opts_knit$set(root.dir = rprojroot::find_rstudio_root_file())
```

# Load libraries

First, we will load the libraries needed for this part of the analysis.

```{r load-libraries, message=FALSE}
sapply(list.files("~/bbvolume/Git/imcRtools/R/", full.names = TRUE), source)

library(data.table)
library(S4Vectors)
library(SingleCellExperiment)
library(devtools)
library(cytomapper)
library(ggplot2)
library(stringr)
```

# Read the data

```{r load data}
sce = readRDS(file = "~/bbvolume/Data/Dermatology/covid_skin_rash/Rout/data/sce_protein.rds")

image_mat <- read.csv(file = "~/bbvolume/Data/Dermatology/covid_skin_rash/processed/cpout/Image.csv")

colour_vector <- metadata(sce)$colour_vectors$layer_1
```

## load the masks

```{r load masks}
all_mask <- loadImages(x = "~/bbvolume/Data/Dermatology/covid_skin_rash/processed/cpout/",pattern = "ilastik_s2_Probabilities_mask.tiff")
```

## add the respective ImageNumber as in the sce object to the all.mask object

```{r match image numbers for masks}


# we extract only the FileNames of the image metadata as they are in the all.masks object
cur_df <- data.frame(cellmask = image_mat$FileName_cellmask,
                     ImageNumber = image_mat$ImageNumber)

# we set the rownames of the extracted data to be equal to the names of all.masks
rownames(cur_df) <- gsub(pattern = ".tiff",replacement = "",image_mat$FileName_cellmask)

# we add the extracted information via mcols in the order of the all.masks object to 
mcols(all_mask) <- cur_df[names(all_mask),]
```

## scale the masks

```{r scale masks}

all_mask <- scaleImages(all_mask,2^16-1)

all_mask
all_mask <- getImages(all_mask,c(unique(sce$ImageNumber)))

```


## now we also load the pixel level data. we will have to do this in batches since loading all stacks at once needs roughly 50 Gb


```{r check celltype assignment}
all_pixel <- loadImages(x = "~/bbvolume/Data/Dermatology/covid_skin_rash/processed/tiffs/",pattern = ".*_full.tiff")
```

## add the respective ImageNumber as in the sce object to the all_pixel object

```{r match image numbers for pixel data}
# we extract only the FileNames of the stacks as they are in the all_pixel object
cur_df <- data.frame(cellmask = image_mat$FileName_FullStack,
                     ImageNumber = image_mat$ImageNumber)

# we set the rownames of the extracted data to be equal to the names of all_pixel
rownames(cur_df) <- gsub(pattern = ".tiff",replacement = "",image_mat$FileName_FullStack)

# we add the extracted information via mcols in the order of the all_pixel object
mcols(all_pixel) <- cur_df[names(all_pixel),]

# scale all_pixel object
all_pixel <- scaleImages(all_pixel,2^16-1)


all_pixel <- getImages(all_pixel,c(unique(sce$ImageNumber)))
```

## channel names in the pixel level data
 
```{r add channel names}
panel_meta <- read.csv(file = "~/bbvolume/Data/Dermatology/covid_skin_rash/Skin_panel.csv", sep= "," )

# remove the E-cadherin marker in channel 139 because it is empty.
panel_meta <- panel_meta[which(! panel_meta$Metal.Tag == "La139"),]

# sort the channel names accordingly to the stack
# extract metal mass
panel_meta$Metal.Number = str_extract(string = panel_meta$Metal.Tag, pattern = "[0-9]+")

# order according to metal mass
panel_meta = panel_meta[order(panel_meta$Metal.Number),]

# assign channel number, matching with the measured channles in cellprofiler
# the channel number should be double-checked with the channel order in the _full.csv file in the tiff folder to see if the order matches (e.g. Y89 can cause a mismatch)
panel_meta$channel = c(1:nrow(panel_meta))

panel_meta$clean_target <- c("HistoneH3", "SMA","CD7", "Filaggrin", "HLADR","CD370", "CD27", "CD16", "CD163", "Langerin", "CD11b", "CD68", "CD31",
                  "CD3", "DC_LAMP", "CD11c", "CD1c", "CD69", "CD134", "GrzB", "CD45RA", "STING", "CD45RO", "FoxP3", "CD20", "CLA",
                  "CD8", "CD57", "Ki67", "DP2","CD40", "CD4", "CD14", "E-cad",  "CD303", "CD206", "DNA1", "DNA2", "MPO")
channelNames(all_pixel) <- panel_meta$clean_target
```

## plot images of the cell types

```{r plot only those cells that have more than 1.5 asinh counts,fig.height=12, fig.width=30}
# this code only runs after the celltypes classification
for(i in 1:16){
png(paste0("~/bbvolume/Data/Dermatology/covid_skin_rash/Rout/cytomapper_images/celltypes_per_image_noscale_no_name_",i,".png"), width = 2500, height = 2500)
plotCells(mask = all_mask[i],sce,cell_id = "CellNumber",img_id = "ImageNumber",colour_by = "celltype_classified",colour = list(celltype_classified = colour_vector),scale_bar =NULL, image_title =NULL)
dev.off()
}
```
